split redir parser
