+++ include/parser.h
@@ -6,7 +6,7 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/24 12:19:42 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/09/29 12:52:34 by tkuramot         ###   ########.fr       */
+/*   Updated: 2023/09/29 16:13:21 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,13 @@
 
 # include "libft.h"
 # include "type.h"
+# include <stdbool.h>
 
 t_ast	*ast_new_node(t_node_type type, t_ast *left, t_ast *right);
 t_ast	*ast_new_node_cmd(t_token *lst);
+bool	expect(t_token *lst, t_token_type type);
+bool	is_redirect(t_token *lst);
+void	arrange_node(t_ast *node);
 void	parse_token(t_context *ctx);
 
 #endif
