+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: tsishika <tsishika@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/19 17:33:13 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/09/29 15:08:06 by tkuramot         ###   ########.fr       */
+/*   Updated: 2023/09/29 15:22:58 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
