+++ src/parser/expect.c
@@ -1,0 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   expect.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: tkuramot <tkuramot@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/29 15:54:07 by tkuramot          #+#    #+#             */
+/*   Updated: 2023/09/29 16:27:38 by tkuramot         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "parser.h"
+#include <stdbool.h>
+
+bool	expect(t_token *lst, t_token_type type)
+{
+	if (!lst)
+		return (false);
+	return (lst->type == type);
+}
