+++ src/parser/parser.c
@@ -6,7 +6,7 @@
 /*   By: tkuramot <tkuramot@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/02 23:57:29 by tkuramot          #+#    #+#             */
-/*   Updated: 2023/09/29 12:57:48 by tkuramot         ###   ########.fr       */
+/*   Updated: 2023/09/29 16:52:29 by tkuramot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,77 +14,6 @@
 #include "ft_string.h"
 #include "lexer.h"
 #include "utils.h"
-
-static bool	expect_pipe(t_token *lst)
-{
-	if (!lst || !lst->next)
-		return (false);
-	return (ft_strcmp(lst->next->word, "|") == 0);
-}
-
-static t_redirect	*init_redir(char *file, t_token_type type)
-{
-	t_redirect	*red;
-
-	red = ft_calloc(1, sizeof (t_redirect));
-	if (!red)
-		fatal_error("malloc");
-	red->file = file;
-	red->type = type;
-	return (red);
-}
-
-static bool		add_redirect(t_list **redirect, t_token **lst)
-{
-	char	*file;
-
-	if (!lst || !*lst || !(*lst)->next)
-		return (false);
-	file = (*lst)->next->word;
-	if ((*lst)->type == TK_REDIR_IN)
-		ft_lstadd_back(redirect, ft_lstnew(init_redir(file, TK_REDIR_IN)));
-	if ((*lst)->type == TK_REDIR_OUT)
-		ft_lstadd_back(redirect, ft_lstnew(
-				init_redir(file, TK_REDIR_OUT)));
-	if ((*lst)->type == TK_REDIR_HEREDOC)
-	{
-		if (ft_strchr(file, '\'') || ft_strchr(file, '\"'))
-			ft_lstadd_back(redirect, ft_lstnew(init_redir(file, TK_REDIR_HEREDOC_Q)));
-		else
-			ft_lstadd_back(redirect, ft_lstnew(init_redir(file, TK_REDIR_HEREDOC)));
-	}
-	if ((*lst)->type == TK_REDIR_APPEND)
-		ft_lstadd_back(redirect, ft_lstnew(
-				init_redir((*lst)->next->word, TK_REDIR_APPEND)));
-	*lst = (*lst)->next->next;
-	return (true);
-}
-
-static void		arrange_node(t_ast *ast)
-{
-	t_token	*lst = ast->lst;
-	t_token	*tmp;
-	t_token	head;
-	t_token	*cur;
-
-	if (!lst)
-		return;
-	tmp = ast->lst;
-	head.next = NULL;
-	cur = &head;
-	while (tmp)
-	{
-		if ((1 << tmp->type) & 0b1111)
-		{
-			add_redirect(&ast->redir_lst, &tmp);
-			continue;
-		}
-		cur->next = token_copy(tmp);
-		cur = cur->next;
-		tmp = tmp->next;
-	}
-	ast->argv = head.next;
-}
 
 static t_ast	*parse_cmd(t_token **lst)
 {
@@ -96,7 +25,7 @@ static t_ast	*parse_cmd(t_token **lst)
 		return (NULL);
 	cmd = *lst;
 	tmp = *lst;
-	while (tmp->next && !expect_pipe(tmp))
+	while (tmp->next && !expect(tmp->next, TK_PIPE))
 		tmp = tmp->next;
 	*lst = tmp->next;
 	tmp->next = NULL;
@@ -108,22 +37,24 @@ static t_ast	*parse_cmd(t_token **lst)
 void	parse_token(t_context *ctx)
 {
 	t_token	*lst;
-	t_ast	*node;
+	t_ast	*left;
+	t_ast	*right;
 
 	lst = ctx->token;
-	node = parse_cmd(&lst);
-	if (!node)
+	left = parse_cmd(&lst);
+	if (!left)
 		fatal_error("malloc");
 	while (true)
 	{
-		if (lst && ft_strcmp(lst->word, "|") == 0)
+		if (lst && expect(lst, TK_PIPE))
 		{
 			lst = lst->next;
-			node = ast_new_node(ND_PIPE, node, parse_cmd(&lst));
+			right = parse_cmd(&lst);
+			left = ast_new_node(ND_PIPE, left, right);
 		}
 		else
 		{
-			ctx->ast = node;
+			ctx->ast = left;
 			return;
 		}
 	}
